Solution to Check MSB of a Number

+++ Check MSB of a Number@@67882a731b4392d8fc1577e9/code.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 int main() {
-    int a;
+    long a;
     scanf("%d",&n);
     if(a | 1){
         printf("Set");
